# services/booking-service/src/main/java/com/ticketing/booking/dto/ConfirmBookingRequest.java
package com.ticketing.booking.dto;

import jakarta.validation.constraints.NotBlank;

public class ConfirmBookingRequest {

    @NotBlank
    private String stripePaymentMethodId;

    public String getStripePaymentMethodId() { return stripePaymentMethodId; }
    public void setStripePaymentMethodId(String stripePaymentMethodId) {
        this.stripePaymentMethodId = stripePaymentMethodId;
    }
}