# services/booking-service/src/main/java/com/ticketing/booking/dto/ErrorResponse.java
package com.ticketing.booking.dto;

import java.time.Instant;

public class ErrorResponse {

    private int status;
    private String error;
    private String message;
    private String path;
    private Instant timestamp = Instant.now();

    public ErrorResponse(int status, String error, String message, String path) {
        this.status = status;
        this.error = error;
        this.message = message;
        this.path = path;
    }

    public int getStatus() { return status; }
    public String getError() { return error; }
    public String getMessage() { return message; }
    public String getPath() { return path; }
    public Instant getTimestamp() { return timestamp; }
}